Test phase 1
Do a simple test suite for the 3 controllers.
The test suite is simple but slow as we truncate for each test suite.

The next step would be to use a database cleaner gem that will hold the
test suite in a transaction.

The complexity here is that there are many databases and each must be
handled separately.